Merge #2: merge test

+++ src/main/java/com/eindopdracht/DJCorner/repositories/FeedbackRepository.java
@@ -3,5 +3,5 @@ package com.eindopdracht.DJCorner.repositories;
 import com.eindopdracht.DJCorner.models.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Integer> {
+public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
 }
